# backend/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Cadrly API
  version: 0.1.0
  description: Приложение для обмена коллекционными карточками
servers:
  - description: Test server
    url: http://localhost:8080/api
...
paths:
...
  /shop/packs:
    get:
      tags: [Shop]
      summary: Получение списка наборов
      description: |
        Возвращает список доступных для покупки наборов карт
      responses:
        200:
          description: Список наборов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pack'
        500:
          $ref: '#/components/responses/ServerError'
  /shop/packs/{pack_id}:
    get:
      tags: [Shop]
      summary: Просмотр содержимого набора
      description: |
        Возвращает подробную информацию о наборе и возможных картах
      parameters:
        - in: path
          name: pack_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Информация о наборе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pack'
        500:
          $ref: '#/components/responses/ServerError'
  /shop/packs/{pack_id}/buy:
    post:
      tags: [Shop]
      summary: Покупка набора карт
      description: |
        Покупает набор карт, если:
        - Пользователь авторизован
        - На балансе достаточно средств
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: pack_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Успешная покупка
          content:
            application/json:
              schema:
                type: object
                properties:
                  receivedCards:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
                  newBalance:
                    type: integer
        401:
          $ref: '#/components/responses/Unauthorized'
        402:
          description: Недостаточно средств
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Недостаточно средств на балансе'
                  requiredAmount:
                    type: integer
                    example: 1000
        500:
          $ref: '#/components/responses/ServerError'
  /shop/packs/{pack_id}/open:
    post:
      tags: [Shop]
      summary: Открытие купленного набора карт
      description: |
        Открывает ранее купленный набор карт, показывая выпавшие карты
        и добавляя их в инвентарь пользователя.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: pack_id
          required: true
          schema:
            type: integer
            example: 1
            description: 'ID набора карт'
      responses:
        200:
          description: Набор успешно открыт
          content:
            application/json:
              schema:
                type: object
                properties:
                  receivedCards:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
                    description: 'Карты, полученные из набора'
                  newCardsAdded:
                    type: boolean
                    example: true
                    description: 'Были ли карты успешно добавлены в инвентарь'
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/ServerError'
  /shop/coins/offers:
    get:
      tags: [Shop]
      summary: Получение предложений по монетам
      description: |
        Возвращает список доступных пакетов монет
      responses:
        200:
          description: Список предложений
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CoinOffer'
        500:
          $ref: '#/components/responses/ServerError'
  /shop/coins/offers/{offer_id}:
    get:
      tags: [Shop]
      summary: Просмотр предложения монет
      description: |
        Возвращает детализацию предложения по монетам
      parameters:
        - in: path
          name: offer_id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Информация о предложении
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoinOffer'
        500:
          $ref: '#/components/responses/ServerError'
  /shop/coins/offers/{offer_id}/purchase:
    post:
      tags: [Shop]
      summary: Покупка монет
      description: |
        Инициирует процесс покупки монет через платежный шлюз
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: offer_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                redirectUrl:
                  type: string
                  format: uri
                  example: 'https://cardly.ru/shop/payment-callback'
              required: [redirectUrl]
      responses:
        200:
          description: Перенаправление на платежный шлюз
          content:
            application/json:
              schema:
                type: object
                properties:
                  paymentUrl:
                    type: string
                    format: uri
        401:
          $ref: '#/components/responses/Unauthorized'
        500:
          $ref: '#/components/responses/ServerError'
  /shop/payments/process:
    post:
      tags: [Shop]
      summary: Обработка платежа
      description: |
        обработка результата платежа от платежного шлюза
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCallback'
      responses:
        200:
          description: Платеж обработан
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transactionId:
                    type: string
        500:
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
...
    Pack:
      type: object
      properties:
        pack_ID:
          type: integer
          example: 401
          description: 'Уникальный идентификатор набора'
        name:
          type: string
          example: 'Стартовый набор'
          description: 'Название набора'
        imageURL:
          type: string
          format: url
          example: 'https://example.com/packs/starter.png'
          description: 'Ссылка на изображение'
        cards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
          description: 'Карты в наборе'
        price:
          type: integer
          example: 1000
          description: 'Цена'
      required:
        - pack_ID
        - name
        - imageURL
        - cards
        - price
    Card:
      type: object
      properties:
        card_ID:
          type: integer
          example: 501
          description: 'Уникальный идентификатор карты'
        name:
          type: string
          example: 'Ледяной феникс'
          description: 'Название карты'
        imageURL:
          type: string
          format: url
          example: 'https://example.com/cards/pheonix.png'
          description: 'Ссылка на изображение'
        description:
          type: string
          example: 'Ледяной Феникс — это мифическое существо, воплощающее силу зимы и вечного обновления. Его перья сверкают как морозный утренний иней, а глаза светятся холодным синим светом.'
          description: 'Описание карты'
        rarity:
          type: string
          enum: ['Обычная', 'Редкая', 'Эпическая', 'Легендарная', 'Уникальная']
          example: 'Редкая'
          description: 'Редкость карты'
        min_price:
          type: integer
          example: 50
          description: 'Минимальная цена - используется для разбора карточки'
        theme:
          type: string
          nullable: true
          example: 'Мифическое существо'
          description: 'Тематика, на которую была сгенерирвана карта (только для уникальных карт)'
        isGenerated:
          type: boolean
          example: false
          description: 'Флаг показывающий сгенерированная ли карта (true для уникальных, false для остальных)'
      required:
        - card_ID
        - name
        - imageURL
        - rarity
        - min_price
        - isGenerated
...
    CoinOffer:
      type: object
      properties:
        offer_id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Стартовый набор'
        coinsAmount:
          type: integer
          example: 100
        price:
          type: number
          format: double
          example: 99.90
        imageUrl:
          type: string
          format: uri
        description:
          type: string
          nullable: true
          example: 'Стартовый набор. Вы можете получить...'
    PaymentCallback:
      type: object
      properties:
        transactionId:
          type: string
        status:
          type: string
          enum: ['Успешно', 'Ошибка', 'В процессе']
        amount:
          type: number
        currency:
          type: string
        offerId:
          type: integer
      required:
        - transactionId
        - status
        - amount
...
  responses:
...
    Unauthorized:
      description: Ошибка аутентификации
    ServerError:
      description: Что-то пошло не так
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'Что-то пошло не так, попробуйте позже'
              code:
                type: string
                example: 'INTERNAL_SERVER_ERROR'
              timestamp:
                type: string
                format: date-time
                example: '2024-05-20T12:00:00Z'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT